requirements.txt
burnysc2>=0.22.3
ollama>=0.4.0
